+++ lib/flutter_smartschool.dart
@@ -31,6 +31,7 @@ export 'src/exceptions.dart';
 export 'src/models/message_models.dart';
 export 'src/models/intradesk_models.dart';
 export 'src/models/notification_models.dart';
+export 'src/models/user_models.dart';
 
 // Services
 export 'src/services/messages_service.dart';